+++ ChangeLog.md
@@ -23,6 +23,7 @@ in the log.
 1. Stop treating zero-item job as failure, to improved the user experience. 
 1. Improved the naming of files being generated in benchmark command, by reversing the digits. 
 Doing so allows the names to not be an alphabetic series, which used to negatively impact the performance on the service side.
+1. Azcopy can now detect when setting a blob tier would be impossible. If azcopy cannot check the destination account type, a new transfer failure status will be set: `TierAvailabilityCheckFailure`
 
 ### Bug fixes